requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==26.1.0
chromedriver-autoinstaller==0.6.4
colorama==0.4.6
exceptiongroup==1.3.1
iniconfig==2.1.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
selenium==3.141.0
tomli==2.4.0
typing_extensions==4.15.0
urllib3==1.26.10
